sub2/backend/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
cPython==0.0.6
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
idna==3.2
joblib==1.0.1
mysqlclient==2.0.3
numpy==1.21.2
pandas==1.3.2
PyJWT==1.7.1
pymongo==3.12.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
scikit-learn==0.22.1
scipy==1.7.1
six==1.16.0
sqlparse==0.4.2
threadpoolctl==2.2.0
urllib3==1.26.6
